# server/docs/swagger.yaml
basePath: /
definitions:
...
  model.AvfOrder:
    properties:
      block:
        description: 区块号
        type: string
      createdAt:
        description: 创建时间
        type: string
      from:
        description: 支付地址
        type: string
      gas:
        description: 手续费
        type: string
      gasPrice:
        description: 手续费价格
        type: string
      id:
        description: 主键ID
        type: integer
      num:
        description: 数量
        type: integer
      number:
        description: 剩余数量
        type: integer
      orderSn:
        description: 订单编号
        type: string
      payTime:
        description: 支付时间
        type: integer
      price:
        description: 价格
        type: integer
      status:
        description: 状态 1-待支付 2-待确认 3-已完成 4-已取消5-已关闭
        type: integer
      to:
        description: 收款地址
        type: string
      txHash:
        description: 交易hash
        type: string
      uid:
        description: 用户ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  model.AvfOrderCard:
    properties:
      cardId:
        description: 卡牌ID
        type: integer
      createTime:
        description: 创建时间
        type: integer
      createdAt:
        description: 创建时间
        type: string
      giveType:
        description: 获得方式 1-抽奖 2-购买
        type: integer
      id:
        description: 主键ID
        type: integer
      level:
        description: 等级 1-N 2-R 3-SR 4-SSR
        type: integer
      orderId:
        description: 订单ID
        type: integer
      star:
        description: 算力值
        type: integer
      status:
        description: 状态 1-正常 2-转让中
        type: integer
      uid:
        description: 用户ID
        type: integer
      updateTime:
        description: 修改时间
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  web_tools.CreateOrderResponse:
    properties:
      order_id:
        description: 订单ID
        type: integer
      order_sn:
        description: 订单号
        type: string
      price:
        description: 订单价格
        type: integer
    type: object
  web_tools.LoginResponse:
    properties:
      id:
        description: 用户ID
        type: integer
      pid:
        description: 上级地址
        type: string
      status:
        description: 状态
        type: boolean
      token:
        description: token
        type: string
      username:
        description: 用户名
        type: string
      wallet_address:
        description: 钱包地址
        type: string
    type: object
  web_tools.OrderDetailResponse:
    properties:
      order_card_list:
        description: 卡牌列表
        items:
          $ref: '#/definitions/model.AvfOrderCard'
        type: array
      order_info:
        $ref: '#/definitions/model.AvfOrder'
        description: 订单详情
    type: object
  web_tools.OrderListResponse:
    properties:
      list:
        description: 订单列表
        items:
          $ref: '#/definitions/model.AvfOrder'
        type: array
      total:
        description: 总条数
        type: integer
    type: object
  web_tools.UserInfo:
    properties:
      avw_balance:
        description: avw余额
        type: number
      ht_balance:
        description: ht余额
        type: number
      id:
        description: 用户ID
        type: integer
      pid:
        description: 上级地址
        type: string
      status:
        description: 状态
        type: boolean
      username:
        description: 用户名
        type: string
      wallet_address:
        description: 钱包地址
        type: string
    type: object
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /web/order/cancelOrder:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: token
        in: header
        name: x-token
        required: true
        type: string
      - description: 订单ID
        in: body
        name: order_id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: status":200,"msg":"成功"}
          schema:
            type: string
      summary: 取消订单
      tags:
      - 前端接口
  /web/order/createOrder:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: x-token
        required: true
        type: string
      - description: 盲盒类型 1-一连抽 2-十连抽
        in: body
        name: type
        required: true
        schema:
          type: string
      - description: 数量 购买盲盒数量
        in: body
        name: num
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_tools.CreateOrderResponse'
      summary: 创建订单
      tags:
      - 前端接口
  /web/order/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: x-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_tools.OrderListResponse'
      summary: 订单列表
      tags:
      - 前端接口
  /web/order/orderDetail:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: token
        in: header
        name: x-token
        required: true
        type: string
      - description: 订单ID
        in: body
        name: order_id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_tools.OrderDetailResponse'
      summary: 订单详情
      tags:
      - 前端接口
  /web/order/payOrder:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: x-token
        required: true
        type: string
      - description: 订单ID
        in: body
        name: order_id
        required: true
        schema:
          type: string
      - description: 发起交易后的事务hash
        in: body
        name: tx_hash
        required: true
        schema:
          type: string
      - description: 发起交易的钱包地址
        in: body
        name: address
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"status":200,"message":"200"}'
          schema:
            type: string
      summary: 支付订单
      tags:
      - 前端接口
  /web/user/getUserInfo:
    get:
      consumes:
      - application/json
      parameters:
      - description: token信息
        in: header
        name: x-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_tools.UserInfo'
      summary: 获取用户信息
      tags:
      - 前端接口
  /web/user/login:
    get:
      consumes:
      - application/json
      parameters:
      - description: 钱包地址
        in: query
        name: wallet_address
        required: true
        type: string
      - description: 上级地址
        in: query
        name: pid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_tools.LoginResponse'
      summary: 登录
      tags:
      - 前端接口
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"